new services

+++ src/main/java/com/pet/parser/model/DataRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface DataRepository extends JpaRepository<DataEntity, Integer> { }
+public interface DataRepository extends JpaRepository<DataEntity, Long> { }
